change file layout
