+++ safepo/algorithms/__init__.py
@@ -12,18 +12,18 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-from safepo.algos.cpo import CPO
-from safepo.algos.cppo_pid import CPPOPid
-from safepo.algos.focops import FOCOPS
-from safepo.algos.ipo import IPO
-from safepo.algos.natural_pg import NPG
-from safepo.algos.p3o import P3O
-from safepo.algos.pcpo import PCPO
-from safepo.algos.policy_gradient import PG
-from safepo.algos.ppo import PPO
-from safepo.algos.ppo_lag import PPO_Lag
-from safepo.algos.trpo import TRPO
-from safepo.algos.trpo_lag import TRPO_Lag
+from safepo.algorithms.cpo import CPO
+from safepo.algorithms.cppo_pid import CPPOPid
+from safepo.algorithms.focops import FOCOPS
+from safepo.algorithms.ipo import IPO
+from safepo.algorithms.natural_pg import NPG
+from safepo.algorithms.p3o import P3O
+from safepo.algorithms.pcpo import PCPO
+from safepo.algorithms.policy_gradient import PG
+from safepo.algorithms.ppo import PPO
+from safepo.algorithms.ppo_lag import PPO_Lag
+from safepo.algorithms.trpo import TRPO
+from safepo.algorithms.trpo_lag import TRPO_Lag
 
 REGISTRY = {
     'pg': PG,
