+++ safepo/algorithms/base.py
