+++ safepo/algorithms/ppo_lag.py
@@ -14,33 +14,40 @@
 # ==============================================================================
 import torch
 
-from safepo.algos.lagrangian_base import Lagrangian
-from safepo.algos.trpo import TRPO
+import safepo.common.mpi_tools as mpi_tools
+from safepo.algorithms.lagrangian_base import Lagrangian
+from safepo.algorithms.policy_gradient import PG
 
 
-class TRPO_Lag(TRPO,Lagrangian):
+class PPO_Lag(PG,Lagrangian):
+    '''
+
+    '''
     def __init__(
             self,
-            algo: str = 'trpo-lag',
-            cost_limit: float = 25.0,
-            lagrangian_multiplier_init: float = 0.001,
-            lambda_lr: float = 0.035,
-            lambda_optimizer: str = 'Adam',
+            algo='ppo-lag',
+            cost_limit=25.,
+            clip=0.2,
+            lagrangian_multiplier_init=0.001,
+            lambda_lr=0.035,
+            lambda_optimizer='Adam',
             use_standardized_reward=True,
             use_standardized_cost=True,
             use_standardized_obs=False,
+            use_reward_scaling=False,
             use_cost_value_function=True,
             use_kl_early_stopping=True,
             **kwargs
-    ):
-        TRPO.__init__(
+        ):
+        PG.__init__(
             self,
             algo=algo,
             use_cost_value_function=use_cost_value_function,
+            use_kl_early_stopping=use_kl_early_stopping,
             use_standardized_reward=use_standardized_reward,
             use_standardized_cost=use_standardized_cost,
             use_standardized_obs=use_standardized_obs,
-            use_kl_early_stopping=use_kl_early_stopping,
+            use_reward_scaling=use_reward_scaling,
             **kwargs
         )
 
@@ -52,22 +59,25 @@ class TRPO_Lag(TRPO,Lagrangian):
             lambda_optimizer=lambda_optimizer
         )
 
+        self.clip = clip
+
     def algorithm_specific_logs(self):
         super().algorithm_specific_logs()
         self.logger.log_tabular('LagrangeMultiplier',
                                 self.lagrangian_multiplier.item())
 
-    def compute_loss_pi(self, data: dict) -> tuple:
+
+    def compute_loss_pi(self, data: dict):
         # Policy loss
         dist, _log_p = self.ac.pi(data['obs'], data['act'])
         ratio = torch.exp(_log_p - data['log_p'])
-
-        loss_pi = -(ratio * data['adv']).mean()
+        ratio_clip = torch.clamp(ratio, 1-self.clip, 1+self.clip)
+        loss_pi = -(torch.min(ratio * data['adv'], ratio_clip * data['adv'])).mean()
         loss_pi -= self.entropy_coef * dist.entropy().mean()
 
         # ensure that lagrange multiplier is positive
-        penalty = torch.clamp_min(self.lagrangian_multiplier,0.0)
-        loss_pi += penalty * (ratio * data['cost_adv']).mean()
+        penalty = self.lambda_range_projection(self.lagrangian_multiplier).item()
+        loss_pi += penalty * ((ratio * data['cost_adv']).mean())
         loss_pi /= (1 + penalty)
 
         # Useful extra info
@@ -81,8 +91,6 @@ class TRPO_Lag(TRPO,Lagrangian):
         raw_data = self.buf.get()
         # pre-process data
         data = self.pre_process_data(raw_data)
-        # sub-sampling accelerates calculations
-        self.fvp_obs = data['obs'][::4]
         # Note that logger already uses MPI statistics across all processes..
         ep_costs = self.logger.get_stats('EpCosts')[0]
         # First update Lagrange multiplier parameter
