+++ safepo/algorithms/trpo.py
@@ -15,7 +15,7 @@
 import torch
 
 import safepo.common.mpi_tools as mpi_tools
-from safepo.algos.natural_pg import NPG
+from safepo.algorithms.natural_pg import NPG
 from safepo.common.utils import get_flat_params_from, set_param_values_to_model
